requirements.txt
wandb~=0.16.0
gdown
numba>0.56
flashlight-text==0.0.4
llvmlite>0.39
matplotlib~=3.8.2
numpy>=1.25
wget
pandas~=2.1.3
Pillow~=10.1.0
torch>=2
torchaudio>=2
kaggle
torchvision>=0.16
tqdm~=4.66.1
inflect~=7.0.0
scipy>=1.11.3
librosa~=0.10.1
scikit-learn>=1.2.2
hydra-core
omegaconf